refactor(messaging/feishu): SOLID + DRY — split god file, extract helpers, reduce duplication
Closes #343

- P0: Split adapter.go (1562→353 lines) into conn.go, handler.go, media.go, ws.go
- P1: Extract WriteCtx 184-line dispatcher into per-event handlers (handleDone,
  handleError, handleToolCall, handleToolResult, handleInteraction, handleStatusEvent,
  handleMessageEvent)
- P2: Extract larkCreateMessage/larkReplyMessage helpers to eliminate 4x Lark IM
  API call duplication across adapter.go, interaction.go, streaming.go
- P3: Extract sendOrReply/sendOrReplyCard/replyOrSend helpers to eliminate 7x
  "if replyToMsgID != '' then reply else send" duplication
- P4: Reorganize StreamingCardController fields with clear section comments
- docs: Add AI Agent section to getting-started.md
